HashMap and HashTable
In Java, HashMap is a part of Java’s collection since Java 1.2. This class is
found in java.util package. It provides the basic implementation of the Map
interface of Java.
HashMap and Hashtable store key and value pairs in a hash table. When using
a Hashtable or HashMap, we specify an object that is used as a key and the value
that you want to be linked to that key. The key is then hashed, and the resulting
hash code is used as the index at which the value is stored within the table.
HashMap is used because it provides features like:
• Fast retrieval
• Efficient storage
• Flexible to use
• Easy to use
• Suitable for large data sets
To use this class and its methods, you need to import java.util.HashMap package
or its superclass.
Create a java code that does the following:
Create a menu that contains and performs the following operations on a
HashMap <Integer, String>:
1. Insert elements ( key ⇒ integer, value⇒string)
2. Update existing element
3. Delete an element
4. Print elements using iterator